Created the bump_f64() and result_bumper_64().
These functions are stupid, dangerous, and stupid dangerous.  However,
there is a use for them; they ensure that the result of a floating
pointer operation is strictly greater than any of their arguments.
This is necessary to ensure that time is always moving forwards.  If
anything goes wrong with your calculations, this is probably the first
place to look.