Now bumping time whenever I set time.
Whenever I call collider's set_time() function, I check to see if the
new time is greater than the current time.  If it isn't I bump the
incoming value up.  This has a knock on effect with the event loop;
it can cause the current time to be after when the next event (or events)
are supposed to execute, which causes a failure.  This commit solves
the problem by taking all events that occurred before the newly bumped
time, and setting their execution time to the newly set time.  This
is probably a bad idea, but I won't know until I run the code (it
passes all internal tests).